test: extract shared WebSocket inspector helper; run concurrently; skip release-ASAN lane
The spawn + URL scan + send/waitForEvent harness was copied from
21654.test.ts; pull it into test/cli/inspect/inspector-ws-helper.ts and
use it from both sites.

Run the three scope-chain tests under describe.concurrent since each
spawns its own debuggee on port 0 in its own tempDir.

Skip on the release-ASAN CI lane (isASAN && !isDebug) where the WebSocket
inspector transport drops connections with close code 1006 (same reason
cli/inspect/inspect.test.ts is quarantined there). The local debug build
also has ASAN enabled but does not hit this, so keep running there.
